test(bpmn): share the simulated-publish helper
